+++ CHANGELOG.md
@@ -5,13 +5,15 @@ _See also: [the official libGDX changelog](https://github.com/libgdx/libgdx/blob
 - **[UPDATE]** Updated to Kotlin 1.9.22.
 - **[UPDATE]** Updated to Gradle 7.6.3.
 - **[CHANGE]** (`ktx-assets`) The `pool` factory method now attempts to reset `Poolable` objects with the default `discard` lambda
-to match the default libGDX pool behavior. Pass a custom function to override it.
+  to match the default libGDX pool behavior. Pass a custom function to override it.
 - **[FEATURE]** (`ktx-math`) Added component deconstructing operators for 2D shapes:
   - `Rectangle`: _x, y, width, height_
   - `Ellipse`: _x, y, width, height_
   - `Circle`: _x, y, radius_
   - `Polygon`: _x, y_
   - `Polyline`: _x, y_
+- **[CHANGE]** (`ktx-scene2d`) `image` factory method can now be called without parameters or with `null` to create
+  an `Image` without a `Drawable`.
 - **[FEATURE]** (`ktx-tiled`) Added `use` extension method to `BatchTiledMapRenderer` that
   automatically calls `beginRender()` and `endRender()`.